refactor(reserve): split positions tab sections
